collections done

+++ src/main/java/dev/rubric/discord/service/SessionService.java
@@ -6,16 +6,16 @@ import org.springframework.stereotype.Service;
 @Service
 public class SessionService {
 
-    public void createUserSession(Integer userId, HttpSession session) {
+    public void createUserSession(Long userId, HttpSession session) {
         session.setAttribute("USER_ID", userId);
     }
 
-    public Integer getUserFromSession(HttpSession session) {
-        return (Integer) session.getAttribute("USER_ID");
+    public Long getUserFromSession(HttpSession session) {
+        return (Long) session.getAttribute("USER_ID");
     }
 
-    public Integer deleteSession(HttpSession session) {
-        Integer userId = (Integer) session.getAttribute("USER_ID");
+    public Long deleteSession(HttpSession session) {
+        Long userId = (Long) session.getAttribute("USER_ID");
         session.invalidate();
         return userId;
     }
